Build GUI & Set up Procedure Structure
Here is the NEW Procedure!
1. Run gui.m
2. Select an eye picture with load button, and it will save selected
image as a testImage.bmp
3. Press compute button to run example.m, which will preduce and save
the result as an outputImage, and to plot sclera, iris, result images at
different axes.

ps. Beware that I've already set up a process structure in example.m.
What we left are 2 well-trained network one for sclera the other for
iris.